Fix Fix Only and Undo Only options
- The "Fix Only" option should not refix a RemapBlend.buf if that file
already exists. Before we did not do this considering the case that the
user's RemapBlend.buf is corrupted and needed to be fixed. But for this
case, the user would not run under the "Fix Only" option

- Add reporting of the upperbound of .ini files that are undoed
- Mark .ini files that are already fixed as fixed instead of not
being fixed
- Add option to clear out the saved text in the `Logger`, in case
someone uses the `BossFixService.fix` multiple times without remembering
the printed text from the last runs of the `BossFixService.fix` function