Added Ayaka Remap
- Ayaka's head and body requires the texture file to be saved as a
BC7 srgb image format. This format does a gamma correction of 1/2.2 ,
causing the texture to be "darker" and more saturate.
However, the Pillow library currently does not support saving .dds
images in different formats due to difficulty porting the
BCn compression encoding algorithm from C/C# to Python. So the current
alternative is to perform to manually perform
a gamma correction of 1/2.2 on the image without saving any metadata to
the header of the image binary.

- Gamma correction filter simplified remap for Ganyu and Ningguang.
Seems like one of the root causes for their remapped heads not matching
the colours of the original was the head being in BC7 Linear format
while the remapped head needs to be in BC7 sRGB format.

- Improved temperature change filter to be closer to the implementation
of Paint.net