Speed up .ini classification by up to 20x times!
Implemented a .ini classification module that answers the following
3 questions:

1. Does the .ini file belong to a mod?
2. What mod does the .ini file belong to?
3. Is the .ini file fixed?

Instead of the old naive implementation for checking many
mod specific regexes over the same .ini string, inspired by DFAs, we
use Aho-Corasick to speed up searching since the problem can be
reduced to searching multiple keywords within a search text.

In Theory:
----------
let n = length of string
let m = # of mods
let k = total length of the keywords to find for the mods

note that usually n >> m and n >> k

so time complexity of old method is O(mn) while time complexity of
new method is O(n + k). Although both can be reduced to be O(n), linear
time, assuming k and m are constant, the constant factor on the
old method would be much higher than the new method with such
assumption.

In Practice:
------------
We implemented our own Aho-Corasick algorithm in pure Python, but this
was too slow, due to Python's infamous high overhead. Luckly, we found
a super fast implementation of the algorithm from pyahocorasick,
which is implemented at the C level.

But this implementation  requires users to have C++ distributables
installed. So by default, the software will use this fast implementation
and fall back to the compatible, but slower implementation if needed.

For the speed comparisons of both old and new methods, see the PR
change of this commit.

The classifier can store state info in another DFA, opening up the ease
to make a classification for WuWa mods.

The classifier also uses some regexes. Since the regex engine is
implemented in C, a standalone regex matching a string is pretty fast.

Lastly, we experimented a bit with concurrency by classifying multiple
.ini files at once. Python's `multiprocessing` is the only way to
use multiple cores on the CPU, but library is succeptible to
key interrupts and processes are a bit heavy weight. We decided to use
`threading`, though in Python, this only achieves concurrency to the
GIL (Global Interpretor Lock). But since we are reading/writing a lot
of files, there still exists a lot of waiting time for context switching
to another thread.

Overall, classification is now a lot faster compared to before
(evident for large mega-merged mods). The current bottleneck would be
the actual parsing/fixing of the mods.