Fix Hutao's hair and optimize texture edits
- Although the original model of CherryHuTao has a blue normal map
for her head, this texture causes some blue outlines around
CherryHuTao's head. We got rid of those outlines by setting the
register for the head's normal map to null.

- Before, some texture edits that do colour replacement are really slow
due to using `PixelFilter`, which explicitely loops over each pixel of
the image in Python. To speed things up, we want to send all the
heavy pixel computation down to the C level. So created new filters for
colour replacement that only uses functions from the `pillow` API to
do the pixel computation.