Add more download modes and fix texture shortening
- Allow more granularity control for the user on how to control the
algorithm for determining what files to download

- Editted texture files require the following information as their
candidate key:
   - Mod To Fix
   - Mod Object
   - Texture Name
   - File Name

`Mod To Fix` and `Mod Object` are already in the original name of the
editted texture, but we do not have the last 2 pieces of info. To not
clutter up the file name to be too long, `Texture Name` and `File Name`
is encoded in a base 64 MD5 encoding that has a integer range from
-65536 to 65536. Repeated hashes are dealt using chaining.
So most of the time, we only use 3 characters to encode some info.