Fix IB handling in HuTao Remap
For some reason, these lines that are typically in a GIMI mod is not
working properly for `HuTao --> CherryHuTao`:

```
[TextureOverride.*IB.*]
hash = ...
handling = skip
drawindexed = auto
...
```

I suspect the error is caused by a model change to CherryHuTao, with
the model change hitting some unhandled edge case for GIMI.

A weird solution would be to duplicate the fixed .ini file and disabling
the section by the name `TextureOverride.*IB.*` for the original .ini
file by setting the key-value pair of `hash`to `null` for those sections

Since `HuTao --> CherryHuTao` originally use a `GIMISplitFixer`, but we
need the multiple file property of a `GIMIMergeFixer`, we relaxed the
restriction on `GIMiMergeFixer` to allow the same source mod object
to merge to many different fixed mod objects. So `GIMIMergeFixer`
became a more powerful version of `GIMISplitFixer`

If this change works, we essentially exposed some of the
internals of GIMI's file search algorithm for how it searches files
within a folder (sinks within the file graph).
We have the following properties:
- The order of the file search within a folder is deterministic
- The order is probably either alphabetical order or reverse
alphabetical order