Simplified and improved asset remplacement
- For assets that require a replacement feature such as hashes or indices,
the structure of the nested dictionary repo may become deeper when
accounting for components. This can simply be an inheritance from the
multi-nested dictionary structure of the [ModDictAssets](https://github.com/nhok0169/Anime-Game-Remap/pull/194)
module.

- Note that with components the old assumption that the assets are
unique is now broken. (eg. YelanTranquil's `BodyA` and `BangA` both have
an index value of 0) So there is a possibility for many different
instances of assets with the same value. To help filter for a specific
instance of an asset to replace, the user can give context to specific
search indices to retrive the correct instance of the asset.